pyproject.toml
[tool.poetry]
name = "repl_python3_beta_gurabot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.6.6"
aiohttp = "^3.7.4"
astroid = "^2.8.2"
async-timeout = "^3.0.1"
attrs = "^21.2.0"
autopep8 = "^1.5.7"
cffi = "^1.15.0"
chardet = "^4.0.0"
"discord.py" = "^1.7.3"
ffmpeg = "^1.4"
idna = "^3.3"
idna_ssl = "^1.1.0"
isort = "^5.9.3"
lazy-object-proxy = "^1.6.0"
mccabe = "^0.6.1"
multidict = "^5.2.0"
platformdirs = "^2.4.0"
pycodestyle = "^2.8.0"
pycparser = "^2.20"
pylint = "^2.11.1"
PyNaCl = "^1.4.0"
six = "^1.16.0"
toml = "^0.10.2"
typed-ast = "^1.4.3"
typing-extensions = "^3.10.0"
yarl = "^1.7.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
